Feat: changing in all the app the int id to long id in player and derivatives

+++ src/main/java/cat/itacademyS5_01/game/dto/GameResponse.java
@@ -1,7 +1,7 @@
 package cat.itacademyS5_01.game.dto;
 
 public record GameResponse(
-        int gameId,
+        Long gameId,
         String playerName,
         int playerScore,
         int bankScore
